backend/requirements.txt
Flask>=2.0
python-dotenv>=0.15
supabase>=1.0
